requirements.txt
# FastAPI + Core
fastapi==0.111.0
uvicorn[standard]==0.25.0
pydantic==2.5.1
python-dotenv==1.0.0
langdetect==1.0.9

# AI / LLMs
openai==1.99.9
google-generativeai==0.8.5
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.40.2
langchain==0.2.0
langgraph==0.3.0
langgraph-sdk==0.1.74
langsmith==0.1.147

# Data Handling
pandas==2.3.1
openpyxl==3.1.5
numpy==1.26.4
PyYAML==6.0.2
requests==2.32.0
beautifulsoup4==4.12.3

# Semantic Search
faiss-cpu==1.7.4

# Torch CPU (no GPU)
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu

# PII Detection (Presidio)
presidio-analyzer==2.2.359
presidio-anonymizer==2.2.359

# SpaCy + Norwegian model
spacy==3.7.5
spacy-lookups-data==1.0.5
nb-core-news-sm==3.7.0

# Testing / Utilities
pytest==8.4.1
tqdm==4.67.1
rich==14.1.0
